# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/wallet/refunds": {
      "get": {
        "operationId": "WalletController_listRefunds",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["payments-wallet"]
      },
      "post": {
        "operationId": "WalletController_proposeRefund",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["payments-wallet"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "paymentIntentId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "amount": {
                    "type": "string",
                    "pattern": "^\\d+(\\.\\d{1,2})?$"
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": ["paymentIntentId", "amount", "reason"]
              }
            }
          }
        }
      }
    },
    "/v1/wallet/refunds/{id}/approve": {
      "patch": {
        "operationId": "WalletController_approveRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["payments-wallet"]
      }
    },
    "/v1/wallet/refunds/{id}/reject": {
      "patch": {
        "operationId": "WalletController_rejectRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["payments-wallet"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": ["reason"]
              }
            }
          }
        }
      }
    },
...
    "/v1/finance/payouts/{id}/approve": {
      "patch": {
        "operationId": "FinanceController_approvePayout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["finance"]
      }
    },
    "/v1/finance/payouts/{id}/reject": {
      "patch": {
        "operationId": "FinanceController_rejectPayout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["finance"]
      }
    },
...
    "/v1/trust-safety/kyc/{id}/propose-decision": {
      "patch": {
        "operationId": "TrustSafetyController_proposeKycDecision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["trust-safety"]
      }
    },
    "/v1/trust-safety/kyc/{id}/confirm-decision": {
      "patch": {
        "operationId": "TrustSafetyController_confirmKycDecision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["trust-safety"]
      }
    },
...
    "/v1/trust-safety/sellers/{sellerId}/suspend": {
      "post": {
        "operationId": "TrustSafetyController_proposeSuspend",
        "parameters": [
          {
            "name": "sellerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["trust-safety"]
      }
    },
    "/v1/trust-safety/sellers/{sellerId}/reinstate": {
      "post": {
        "operationId": "TrustSafetyController_proposeReinstate",
        "parameters": [
          {
            "name": "sellerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["trust-safety"]
      }
    },
    "/v1/trust-safety/suspension-requests/{id}/confirm": {
      "patch": {
        "operationId": "TrustSafetyController_confirmSuspension",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["trust-safety"]
      }
    },
    "/v1/trust-safety/suspension-requests/{id}/reject": {
      "patch": {
        "operationId": "TrustSafetyController_rejectSuspension",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["trust-safety"]
      }
    },
...
    "/v1/trust-safety/disputes/{id}": {
      "get": {
        "operationId": "TrustSafetyController_getDispute",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["trust-safety"]
      }
    },
    "/v1/trust-safety/disputes/{id}/comments": {
      "post": {
        "operationId": "TrustSafetyController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["trust-safety"]
      }
    },
    "/v1/trust-safety/disputes/{id}/resolve": {
      "patch": {
        "operationId": "TrustSafetyController_resolveDispute",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["trust-safety"]
      }
    },
...
  },
  "info": {
    "title": "Nova API",
    "description": "Phase 1 slice — Identity, Catalog, Cart & Checkout, Orders. See backend/docs/02-api-standards.md.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}